fastgen-api/app/data/templates/auth-ad/requirements.txt
aiomysql==0.1.0
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
cffi==1.15.0
click==8.1.2
cryptography==36.0.2
databases==0.5.5
ecdsa==0.17.0
fastapi==0.75.1
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
ldap3==2.9.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
PyMySQL==1.0.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.2
